:rocket: release 0.3.0
