+++ docs/api/configs.md
@@ -1,0 +1,21 @@
+# Configuration API Reference
+
+Configuration classes for Ray execution and data processing options.
+
+---
+
+::: dagster_ray.configs.Lifecycle
+    options:
+      members:
+        - "create"
+        - "wait"
+        - "connect"
+        - "cleanup"
+
+::: dagster_ray.configs.RayDataExecutionOptions
+    options:
+      members: true
+
+::: dagster_ray.configs.ExecutionOptionsConfig
+    options:
+      members: true
