+++ justfile
@@ -10,11 +10,14 @@ publish-dev:
     uv run --with hatch hatch clean
     echo '__version__ = "0.0.0"  # managed by hatch' > src/dagster_ray/_version.py
 
-serve-docs:
+docs-build:
+    uv run --group docs mkdocs build --clean --strict
+
+docs-serve:
     uv run --group docs mkdocs serve
 
-publish-docs:
-    uv run --group docs --all-extras mike deploy --push --update-aliases $(uv run dunamai from any --style pep440) dev
+docs-publish:
+    uv run --group docs --all-extras mike deploy --push --update-aliases $(uv run dunamai from any --style pep440)
 
 ruff:
     uv run --no-sync ruff check --fix --exit-zero
