+++ src/dagster_ray/core/executor.py
@@ -27,9 +27,9 @@ from dagster._utils.merger import merge_dicts
 from packaging.version import Version
 from pydantic import Field
 
-from dagster_ray.config import RayExecutionConfig, RayJobSubmissionClientConfig
+from dagster_ray.configs import RayExecutionConfig, RayJobSubmissionClientConfig
+from dagster_ray.core.run_launcher import RayRunLauncher
 from dagster_ray.kuberay.utils import get_k8s_object_name
-from dagster_ray.run_launcher import RayRunLauncher
 from dagster_ray.utils import resolve_env_vars_list
 
 if TYPE_CHECKING:
