+++ src/dagster_ray/kuberay/resources/raycluster.py
@@ -2,7 +2,8 @@ import dagster as dg
 from pydantic import Field, PrivateAttr
 from typing_extensions import override
 
-from dagster_ray._base.resources import BaseRayResource, Lifecycle
+from dagster_ray._base.resources import RayResource
+from dagster_ray.configs import Lifecycle
 from dagster_ray.kuberay.client import RayClusterClient
 from dagster_ray.kuberay.client.base import load_kubeconfig
 from dagster_ray.kuberay.configs import RayClusterConfig
@@ -23,7 +24,7 @@ class KubeRayClusterClientResource(dg.ConfigurableResource[RayClusterClient]):
         return RayClusterClient(kube_context=self.kube_context, kube_config=self.kube_config)
 
 
-class KubeRayCluster(BaseKubeRayResourceConfig, BaseRayResource):
+class KubeRayCluster(BaseKubeRayResourceConfig, RayResource):
     """
     Provides a `RayCluster` for Dagster steps.
